+++ backend/src/main/java/app/repository/LibraryRepository.java
@@ -4,5 +4,5 @@ import app.domain.Book;
 import app.domain.Library;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LibraryRepository extends JpaRepository<Library, Integer> {
+public interface LibraryRepository extends JpaRepository<Library, Long> {
 }
